// angular-frontend/src/app/core/models/MailSuffix.ts
/**
 * Copyright Open Logistics Foundation
 *
 * Licensed under the Open Logistics Foundation License 1.3.
 * For details on the licensing terms, see the LICENSE file.
 * SPDX-License-Identifier: OLFL-1.3
 */

export interface MailSuffix {
    id: number;
    mailSuffix: string;
}

export interface ExportImportMailSuffix {
    mailSuffix: string;
}